2001-09-06  Alex Angas  <wgd@adelaide.on.net>
	* src/sound/soundserver.c: Did some very initial commenting.

	* src/include/midiout.h: Added #include <sci_memory.h>.

	* src/win32/sciv.dsp, src/win32/fsci_dll.dsp: Changed SDL directory to
	generic name.

	* src/engine/kernel.c (kGetTime): Added check of return value for
	timeEndPeriod().

	* src/scicore/tools.c (sci_gettime): Removed extra timeBeginPeriod
	call and added check of return value for timeEndPeriod().

	* src/engine/kernel.c; src/gfx/operations.c;
	src/gfx/drivers/xlib_driver.c, sdl_driver.c, dd_driver.cpp,
	ggi_driver.c; src/include/sci_conf.h, resource.h, gfx_operations.h,
	gfx_driver.h, engine.h, soundserver.h; src/scicore/tools.c;
	src/sound/soundserver.c, sound.c, sfx_save.c, sfx_save.cfsml: Made all
	types used for timers consistent (changed from ints to longs).
	Reformatted with indenting.

+++ src/gfx/drivers/dd_driver.cpp
@@ -731,7 +731,7 @@ dd_get_event(gfx_driver_t *drv)
 }
 
 static void
-MsgWait (int WaitTime)
+MsgWait (long WaitTime)
 {
 	DWORD dwRet=0;
 	long dwWait;
@@ -762,7 +762,7 @@ MsgWait (int WaitTime)
 }
 
 int
-dd_usleep(gfx_driver_t* drv, int usec)
+dd_usleep(gfx_driver_t* drv, long usec)
 {
 	if (usec < 1000)
 		Sleep(usec / 1000);
